printf

+++ ft_printf.h
@@ -15,12 +15,12 @@
 
 # include <unistd.h>
 # include <stdarg.h>
-# include "./libft/libft.h"
+//# include "./libft/libft.h"
 
 int		ft_printf(const char *format, ...);
 int		if_per(const char *start, va_list ap, int count);
 int		put_hex(unsigned long nb, int upper);
-int		put_pointer(unsigned long long p);
+int		put_pointer(unsigned long p);
 int		put_str(char *s);
 int		put_c(int c);
 int		put_unsign_int(unsigned int nb);
